requirements.txt
pandas=2.2.3
python-dotenv=1.0.1
pymongo=4.10.1
psycopg2=2.9.10